Fix integers in sample

+++ modules/gapi/samples/onevpl_infer_single_roi.cpp
@@ -250,8 +250,8 @@ int main(int argc, char *argv[]) {
     std::string file_path = cmd.get<std::string>("input");
     const std::string output = cmd.get<std::string>("output");
     const auto face_model_path = cmd.get<std::string>("facem");
-    const auto streaming_queue_capacity = cmd.get<size_t>("streaming_queue_capacity");
-    const auto source_queue_capacity = cmd.get<size_t>("source_queue_capacity");
+    const auto streaming_queue_capacity = cmd.get<uint64_t>("streaming_queue_capacity");
+    const auto source_queue_capacity = cmd.get<uint64_t>("source_queue_capacity");
 
     // check ouput file extension
     if (!output.empty()) {
